Move patternlab style files to app/patternlab. Move app custom files to app/custom. Changes to sync with last version of ng-patternlab.
