+++ app/custom/_style.scss
@@ -1,0 +1,5 @@
+/* app styles */
+@import 'atoms/style';
+@import 'molecules/style';
+@import 'organisms/style';
+@import 'templates/style';
