+++ app/custom/atoms/_style.scss
@@ -5,7 +5,6 @@
 @import 'base/media';
 @import 'base/headings';
 @import 'base/links';
-@import 'base/forms';
 @import 'base/lists';
 @import 'base/icomoon';
 @import 'base/text';
@@ -21,5 +20,3 @@
 @import 'ui/animations';
 
 @import 'img/figure-responsive';
-
-@import 'ln-patternlab/style';
