+++ app/custom/atoms/base/_grid-settings.scss
@@ -1,4 +1,4 @@
-@import '../../../bower_components/neat/app/assets/stylesheets/neat-helpers';
+@import '../../../../bower_components/neat/app/assets/stylesheets/neat-helpers';
 
 // Adding a Visual Grid while you develop. Set to 'true.'
 $visual-grid: false;
