+++ app/custom/atoms/img/_figure-responsive.scss
