+++ app/patternlab/_style.scss
@@ -1,0 +1,5 @@
+/* patternlab styles */
+@import 'atoms/style';
+@import 'molecules/style';
+@import 'organisms/style';
+@import 'templates/style';
