+++ app/patternlab/atoms/_style.scss
@@ -1,0 +1,1 @@
+/* patternlab atoms styles */
