+++ app/patternlab/molecules/_ln-m-widget-recent.scss
@@ -1,0 +1,19 @@
+.ln-m-widget-recent {
+  ul {
+    padding: 0;
+  }
+
+  li {
+    list-style-type: none;
+  }
+
+  h4 {
+    text-transform: uppercase;
+  }
+
+  a {
+    display: block;
+    margin-bottom: 3em;
+    position: relative;
+  }
+}
