+++ app/patternlab/molecules/_style.scss
@@ -1,5 +1,7 @@
+/* patternlab molecules styles */
 @import 'ln-m-widget-contact';
 @import 'ln-m-widget-downloads';
 @import 'ln-m-widget-menu';
 @import 'ln-m-widget-preview';
+@import 'ln-m-widget-recent';
 @import 'ln-m-widget-social';
