+++ app/patternlab/organisms/_style.scss
@@ -1,0 +1,1 @@
+/* patternlab organisms styles */
