+++ app/patternlab/templates/_style.scss
@@ -1,0 +1,1 @@
+/* patternlab templates styles */
