+++ app/style.scss
@@ -14,8 +14,6 @@
 @import '../bower_components/fontSize.scss/font-size';
 @import '../bower_components/wp.scss/wp';
 
-/* app styles */
-@import 'atoms/style';
-@import 'molecules/style';
-@import 'organisms/style';
-@import 'templates/style';
+/* styles */
+@import 'custom/style';
+@import 'patternlab/style';
